requirements.txt
oauthlib>=3.2.2
requests>=2.31.0
python-dotenv>=1.0.0
requests_oauthlib>=1.3.1